+++ src/main/java/com/syncleus/ferma/framefactories/annotation/AnnotationFrameFactory.java
@@ -32,94 +32,44 @@ import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.*;
 
-public class AnnotationFrameFactory implements FrameFactory {
-
-    private final Map<Class<? extends Annotation>, MethodHandler> methodHandlers = new HashMap<>();
-    private final Map<Class, Class> constructedClassCache = new HashMap<>();
-    private final ReflectionCache reflectionCache;
+public class AnnotationFrameFactory extends AbstractAnnotationFrameFactory {
 
     public AnnotationFrameFactory(final ReflectionCache reflectionCache) {
-        this.reflectionCache = reflectionCache;
+        super(reflectionCache, collectHandlers(null));
+    }
+
+    /**
+     * Subclasses can use this constructor to add additional custom method handlers.
+     *
+     * @param reflectionCache The reflection cache used to inspect annotations.
+     * @param handlers The handlers used to generate new annotation support.
+     */
+    protected AnnotationFrameFactory(final ReflectionCache reflectionCache, Set<MethodHandler> handlers) {
+        super(reflectionCache, collectHandlers(handlers));
+    }
+
+    private static final Set<MethodHandler> collectHandlers(Set<MethodHandler> additionalHandlers) {
+        final Set<MethodHandler> methodHandlers = new HashSet<>();
 
         final PropertyMethodHandler propertyHandler = new PropertyMethodHandler();
-        methodHandlers.put(propertyHandler.getAnnotationType(), propertyHandler);
+        methodHandlers.add(propertyHandler);
 
         final InVertexMethodHandler inVertexHandler = new InVertexMethodHandler();
-        methodHandlers.put(inVertexHandler.getAnnotationType(), inVertexHandler);
+        methodHandlers.add(inVertexHandler);
 
         final OutVertexMethodHandler outVertexHandler = new OutVertexMethodHandler();
-        methodHandlers.put(outVertexHandler.getAnnotationType(), outVertexHandler);
+        methodHandlers.add(outVertexHandler);
 
         final AdjacencyMethodHandler adjacencyHandler = new AdjacencyMethodHandler();
-        methodHandlers.put(adjacencyHandler.getAnnotationType(), adjacencyHandler);
+        methodHandlers.add(adjacencyHandler);
 
         final IncidenceMethodHandler incidenceHandler = new IncidenceMethodHandler();
-        methodHandlers.put(incidenceHandler.getAnnotationType(), incidenceHandler);
+        methodHandlers.add(incidenceHandler);
+
+        if(additionalHandlers != null)
+            methodHandlers.addAll(additionalHandlers);
+
+        return methodHandlers;
     }
 
-    @Override
-    public <T> T create(final Element e, final Class<T> kind) {
-
-        Class<? extends T> resolvedKind = kind;
-        if (isAbstract(resolvedKind))
-            resolvedKind = constructClass(e, kind);
-        try {
-            final T object = resolvedKind.newInstance();
-            if (object instanceof CachesReflection)
-                ((CachesReflection) object).setReflectionCache(this.reflectionCache);
-            return object;
-        }
-        catch (final InstantiationException | IllegalAccessException caught) {
-            throw new IllegalArgumentException("kind could not be instantiated", caught);
-        }
-    }
-
-    private static boolean isAbstract(final Class<?> clazz) {
-        return Modifier.isAbstract(clazz.getModifiers());
-    }
-
-    private static boolean isAbstract(final Method method) {
-        return Modifier.isAbstract(method.getModifiers());
-    }
-
-    private <E> Class<? extends E> constructClass(final Element element, final Class<E> clazz) {
-        Class constructedClass = constructedClassCache.get(clazz);
-        if (constructedClass != null)
-            return constructedClass;
-
-        DynamicType.Builder<? extends E> classBuilder;
-        if (clazz.isInterface())
-            if (element instanceof Vertex)
-                classBuilder = (DynamicType.Builder<? extends E>) new ByteBuddy().subclass(AbstractVertexFrame.class).implement(clazz);
-            else if (element instanceof Edge)
-                classBuilder = (DynamicType.Builder<? extends E>) new ByteBuddy().subclass(AbstractEdgeFrame.class).implement(clazz);
-            else
-                throw new IllegalStateException("class is neither an Edge or a vertex!");
-        else {
-            if (element instanceof Vertex && !VertexFrame.class.isAssignableFrom(clazz))
-                throw new IllegalStateException(clazz.getName() + " Class is not a type of VertexFrame");
-            if (element instanceof Edge && !EdgeFrame.class.isAssignableFrom(clazz))
-                throw new IllegalStateException(clazz.getName() + " Class is not a type of EdgeFrame");
-            classBuilder = new ByteBuddy().subclass(clazz);
-        }
-
-        classBuilder = classBuilder.defineField("reflectionCache", ReflectionCache.class, Visibility.PRIVATE, FieldManifestation.PLAIN).implement(CachesReflection.class).intercept(FieldAccessor.
-              ofBeanProperty());
-
-        //try and construct any abstract methods that are left
-        for (final Method method : clazz.getMethods())
-            if (isAbstract(method))
-                annotation_loop:
-                for (final Annotation annotation : method.getAnnotations()) {
-                    final MethodHandler handler = methodHandlers.get(annotation.annotationType());
-                    if (handler != null) {
-                        classBuilder = handler.processMethod(classBuilder, method, annotation);
-                        break;
-                    }
-                }
-
-        constructedClass = classBuilder.make().load(AnnotationFrameFactory.class.getClassLoader(), ClassLoadingStrategy.Default.WRAPPER).getLoaded();
-        this.constructedClassCache.put(clazz, constructedClass);
-        return constructedClass;
-    }
 }